Install webpack
