+++ public/index.html
@@ -9,7 +9,5 @@
 <body>
   <div id="app"></div>
 </body>
-<script src="https://unpkg.com/react@16.0.0/umd/react.development.js"></script>
-<script src="https://unpkg.com/react-dom@16.0.0/umd/react-dom.development.js"></script>
-<script src="/scripts/app.js"></script>
+<script src="/bundle.js"></script>
 </html>
